Refactoring: Separate modal from index.ts
